moved xmlparser back to wml so that namespaces would be localized; updated waterml doc to external cuahsi docs; removed duplicated dict_union
